AI-Agent-for-Document-QA/requirements.txt
langchain
langchain-experimental
langchain-huggingface
langchain-community
langchain-openai
chromadb

python-dotenv
streamlit

transformers
accelerate
huggingface-hub

pymupdf
pdf2image
Pillow

python-docx

# Utilidades adicionales comunes
tqdm